break up crm into files
